issue92Test and issue99Test reorganized
