+++ mockk/jvm/src/test/kotlin/io/mockk/it/MockStaticMethodTest.kt
@@ -1,14 +1,33 @@
-package io.mockk.gh
+package io.mockk.it
 
 import io.mockk.every
 import io.mockk.mockkStatic
 import io.mockk.unmockkStatic
+import org.junit.Test
 import java.time.Instant
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
 
-class Issue99Test {
+/**
+ * Tests related to github issues #92 | #99
+ */
+class MockStaticMethodTest {
+    /**
+     * github issue #92
+     */
+    @Test
+    fun test() {
+        val epochSeconds = 123L
+        mockkStatic(Instant::class)
+        every { Instant.now().epochSecond } returns epochSeconds
+
+        assertEquals(123L, Instant.now().epochSecond)
+    }
+
+    /**
+     * github issue #99
+     */
     @Test
     fun unmockStatic_unmocksStaticMocks() {
         mockkStatic(Instant::class)
